refactor : ColorBoard 삭제 후, ChessBoard 로 역할 위임
