+++ src/test/java/domain/chessboard/ChessBoardTest.java
@@ -1,7 +1,8 @@
 package domain.chessboard;
 
-import domain.chessboard.ChessBoard;
-import domain.chessboard.Rank;
+import domain.coordinate.PositionFactory;
+import domain.piece.Color;
+import domain.piece.Pawn;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -22,6 +23,117 @@ public class ChessBoardTest {
                 .sum();
 
         assertThat(sum).isEqualTo(64);
+    }
+
+    /*
+    RNBQKBNR
+    PPPPPPPP
+    ........
+    ........
+    ........
+    ........
+    pppppppp
+    rnbqkbnr
+    */
+    @Test
+    @DisplayName("초기 상태에는 양쪽 모두 38점이다.")
+    void scoreIsThirtyEight() {
+        // Given
+        ChessBoard chessBoard = ChessBoard.generate();
+
+        // When
+        double blackScore = chessBoard.calculateColorScore(Color.BLACK);
+        double whiteScore = chessBoard.calculateColorScore(Color.WHITE);
+
+        // Then
+        assertThat(blackScore).isEqualTo(38);
+        assertThat(whiteScore).isEqualTo(38);
+    }
+
+    /*
+    RNBQKBNR
+    PPPPPPPP
+    ........
+    ........
+    p.......
+    p.......
+    pppppppp
+    rnbqkbnr
+     */
+    @Test
+    @DisplayName("pawn이 세로로 위치해있으면 0.5점으로 판별된다, 즉, 39점이다. (WHITE)")
+    void scoreIsThirtyNine() {
+        // Given
+        ChessBoard chessBoard = ChessBoard.generate();
+        chessBoard.findSquare(PositionFactory.createPosition("a3"))
+                .bePiece(new Square(new Pawn(Color.WHITE)));
+        chessBoard.findSquare(PositionFactory.createPosition("a4"))
+                .bePiece(new Square(new Pawn(Color.WHITE)));
+
+        // When&Then
+        assertThat(chessBoard.calculateColorScore(Color.WHITE)).isEqualTo(38.5);
+    }
+
+    /*
+    RNBQKBNR
+    PPPPPPPP
+    ........
+    ........
+    ........
+    ........
+    pppppppp
+    rnbqkbnr
+    */
+    @Test
+    @DisplayName("둘다 King 이 있는 경우")
+    void allExistKing() {
+        // Given
+        ChessBoard chessBoard = ChessBoard.generate();
+
+        // When&Then
+        assertThat(chessBoard.isCheckMate()).isFalse();
+    }
+
+    /*
+    RNBQ.BNR
+    PPPPPPPP
+    ........
+    ........
+    ........
+    ........
+    pppppppp
+    rnbqkbnr
+    */
+    @Test
+    @DisplayName("black 이 king 이 없는 경우")
+    void blackNotExistKing() {
+        // Given
+        ChessBoard chessBoard = ChessBoard.generate();
+        chessBoard.findSquare(PositionFactory.createPosition("e8")).beEmpty();
+
+        // When&Then
+        assertThat(chessBoard.isCheckMate()).isTrue();
+    }
+
+    /*
+    RNBQKBNR
+    PPPPPPPP
+    ........
+    ........
+    ........
+    ........
+    pppppppp
+    rnbq.bnr
+    */
+    @Test
+    @DisplayName("White 가 king 이 없는 경우")
+    void whiteNotExistKing() {
+        // Given
+        ChessBoard chessBoard = ChessBoard.generate();
+        chessBoard.findSquare(PositionFactory.createPosition("e1")).beEmpty();
+
+        // When&Then
+        assertThat(chessBoard.isCheckMate()).isTrue();
     }
 
 }
